create application config
